Créer un composant dédié pour la création de l'achat
